refactor: Pawn 클래스와 Board 클래스를 chess 패키지로 분리
