+++ src/test/java/chess/BoardTest.java
@@ -1,4 +1,4 @@
-package softeer2nd;
+package chess;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
